Move NullPopulator to a separate cs file.
