+++ src/ContentRepository/ContentRepository.csproj
@@ -163,6 +163,7 @@
     <Compile Include="Configuration\Services.cs" />
     <Compile Include="Configuration\Versioning.cs" />
     <Compile Include="ISnComponent.cs" />
+    <Compile Include="Search\Indexing\NullPopulator.cs" />
     <Compile Include="SnComponent.cs" />
     <Compile Include="Linq\CaseInsensitiveFilterVisitor.cs" />
     <Compile Include="Packaging\Dependency.cs" />
